requirements.txt
pandas==2.2.3
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0